refactor : 물품 구매 관련 금액은 Long -> Integer로 변경

+++ src/main/java/jpabook/jpashop/domain/user/Account.java
@@ -42,15 +42,15 @@ public class Account extends BaseEntity {
     }
 
 
-    public void withdraw(Long amount) throws InvalidBalanceValueException {
+    public void withdraw(int amount) throws InvalidBalanceValueException {
         if(balance <= amount){
             throw new InvalidBalanceValueException(AccountExceptionMessages.NEGATIVE_ACCOUNT_BALANCE.getMessage());
         }
         balance -= amount;
     }
 
-    public void deposit(long amount) throws InvalidBalanceValueException {
-        if(Long.MAX_VALUE - balance < amount){
+    public void deposit(int amount) throws InvalidBalanceValueException {
+        if(Integer.MAX_VALUE - balance < amount){
             throw new InvalidBalanceValueException(AccountExceptionMessages.BALANCE_OVERFLOW.getMessage());
         }
         balance += amount;
